- Made the db collection that holds the results of the feature extraction and the ml for the active directory data, more generic, in such a way that other classifiers results will be added to the same collection. - Added feature list to the db collection that holds the results of the feature extraction and the ml. Before every feature was an element of it own. now the same information is also hold by the list. this duplication will help for different requests. - Added the score results to the User collection - REST: adding user scores and user score history rest api. - REST: returning the scores as int and not double. - REST: returning the feature value as a string. - REST: changing some of the attribute names in the user details. - For now permiting all url path with out authentication.

+++ fortscale/fortscale-services/src/main/java/fortscale/services/IUserScore.java
@@ -3,6 +3,6 @@ package fortscale.services;
 public interface IUserScore {
 	public String getId();
 	public String getName();
-	public Double getScore();
-	public Double getAvgScore();
+	public int getScore();
+	public int getAvgScore();
 }
